chore(web): collapse settings hooks and dead wrappers
