+++ apps/web/package.json
@@ -21,7 +21,6 @@
     "@streamdown/mermaid": "^1.0.2",
     "@tanstack/react-query": "^5.100.14",
     "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "hashvatar": "^0.1.2",
     "hugeicons-react": "^0.4.0",
