+++ apps/web/src/hooks/use-worker-actions.ts
@@ -2,35 +2,25 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { client } from "@/lib/client";
 import { queryKeys } from "@/lib/query-keys";
 
-export function useStartWorker() {
+function useWorkerMutation(mutationFn: (name: string) => Promise<unknown>) {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (name: string) => client.startWorker(name),
+    mutationFn,
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: queryKeys.systemStatus });
     },
   });
+}
+
+export function useStartWorker() {
+  return useWorkerMutation((name) => client.startWorker(name));
 }
 
 export function useStopWorker() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (name: string) => client.stopWorker(name),
-    onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: queryKeys.systemStatus });
-    },
-  });
+  return useWorkerMutation((name) => client.stopWorker(name));
 }
 
 export function useRestartWorker() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (name: string) => client.restartWorker(name),
-    onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: queryKeys.systemStatus });
-    },
-  });
+  return useWorkerMutation((name) => client.restartWorker(name));
 }
