+++ apps/web/src/pages/files/files-artifact-views.tsx
@@ -2,10 +2,9 @@ import type { ArtifactFile } from "@nakama/core/contract";
 import { Button } from "@/components/ui/button";
 import { formatError } from "@/lib/client";
 import type { FilesViewMode } from "@/lib/files-page.shared";
+import { ArtifactFolderCard } from "@/pages/files/files-artifact-folder-card";
 import type { ArtifactFolderEntry } from "@/pages/files/files-artifact-folders";
-import { ArtifactGridSkeleton } from "@/pages/files/files-artifact-grid-skeleton";
-import { ArtifactGridView } from "@/pages/files/files-artifact-grid-view";
-import { ArtifactListSkeleton } from "@/pages/files/files-artifact-list-skeleton";
+import { ArtifactGridCard } from "@/pages/files/files-artifact-grid-card";
 import { ArtifactListView } from "@/pages/files/files-artifact-list-view";
 
 type FilesArtifactPagination = {
@@ -13,6 +12,94 @@ type FilesArtifactPagination = {
   onShowMore: () => void;
   remainingCount: number;
 };
+
+function ArtifactGridSkeleton() {
+  return (
+    <ul
+      aria-hidden
+      className="grid grid-cols-[repeat(auto-fill,minmax(11rem,1fr))] gap-3 p-3"
+    >
+      {Array.from({ length: 6 }).map((_, index) => (
+        <li
+          className="overflow-hidden rounded-md border border-border"
+          key={`artifact-grid-skeleton-${index}`}
+        >
+          <div className="skeleton-shimmer aspect-[4/3] w-full" />
+          <div className="space-y-2 p-3">
+            <div className="skeleton-shimmer h-4 w-3/4 rounded" />
+            <div className="skeleton-shimmer h-3 w-1/2 rounded" />
+            <div className="skeleton-shimmer h-3 w-2/3 rounded" />
+          </div>
+        </li>
+      ))}
+    </ul>
+  );
+}
+
+function ArtifactListSkeleton() {
+  return (
+    <ul aria-hidden className="divide-y divide-border">
+      {Array.from({ length: 6 }).map((_, index) => (
+        <li
+          className="flex items-center justify-between gap-3 px-4 py-3"
+          key={`artifact-skeleton-${index}`}
+        >
+          <div className="flex min-w-0 items-start gap-3">
+            <div className="skeleton-shimmer mt-0.5 size-4 shrink-0 rounded" />
+            <div className="min-w-0 space-y-1.5">
+              <div className="skeleton-shimmer h-4 w-48 max-w-full rounded" />
+              <div className="skeleton-shimmer h-3 w-64 max-w-full rounded" />
+            </div>
+          </div>
+          <div className="flex shrink-0 items-center gap-3">
+            <div className="skeleton-shimmer size-8 rounded-md" />
+            <div className="skeleton-shimmer size-8 rounded-md" />
+          </div>
+        </li>
+      ))}
+    </ul>
+  );
+}
+
+function ArtifactGridView({
+  profileId,
+  folders,
+  artifacts,
+  deletePending,
+  showFullPath,
+  onDelete,
+  onOpenFolder,
+}: {
+  profileId: string;
+  folders: ArtifactFolderEntry[];
+  artifacts: ArtifactFile[];
+  deletePending: boolean;
+  showFullPath: boolean;
+  onDelete: (artifact: ArtifactFile) => void;
+  onOpenFolder: (prefix: string) => void;
+}) {
+  return (
+    <ul className="grid grid-cols-[repeat(auto-fill,minmax(11rem,1fr))] gap-3">
+      {folders.map((folder) => (
+        <ArtifactFolderCard
+          folder={folder}
+          key={folder.prefix}
+          onOpen={onOpenFolder}
+        />
+      ))}
+      {artifacts.map((artifact) => (
+        <ArtifactGridCard
+          artifact={artifact}
+          deletePending={deletePending}
+          key={artifact.filename}
+          onDelete={() => onDelete(artifact)}
+          profileId={profileId}
+          showFullPath={showFullPath}
+        />
+      ))}
+    </ul>
+  );
+}
 
 function ShowMoreArtifactsButton({
   loadingMore,
